+++ R/surv-brier.R
@@ -1,21 +1,21 @@
 #' Brier score for right censored data
 #'
 #' Compute the time-dependent Brier score for right censored data, which is the
-#' mean squared error at time point `eval_time`.
+#' mean squared error at time point `.eval_time`.
 #'
 #' @family dynamic survival metrics
 #' @templateVar fn brier_survival
 #' @template return-dynamic-survival
 #' @details
 #'
-#' This method automatically groups by the `eval_time` argument.
+#' This method automatically groups by the `.eval_time` argument.
 #'
 #' Smaller values of the score are associated with better model performance.
 #'
 #' @inheritParams pr_auc
 #'
 #' @param data A `data.frame` containing the columns specified by `truth` and
-#' `estimate`.
+#' `...`.
 #'
 #' @param truth The column identifier for the true survival result (that
 #' is created using [survival::Surv()].). This should be an unquoted column name
@@ -23,25 +23,12 @@
 #' [quasiquotation][rlang::quasiquotation] (you can unquote column names). For
 #' `_vec()` functions, an [survival::Surv()] object.
 #'
-#' @param estimate The column identifier for the survival probabilities (that should be
-#' numeric). This should be an unquoted column name although
-#' this argument is passed by expression and supports
-#'  [quasiquotation][rlang::quasiquotation] (you can unquote column names). For
-#' `_vec()` functions, a numeric vector.
-#'
-#' @param censoring_weights The column identifier for censoring weights (that should
-#' be numeric). This should be an unquoted column name although
-#' this argument is passed by expression and supports
+#' @param ... The column identifier for the survival probabilities this
+#' should be a list column of data.frames corresponding to the output given when
+#' predicting with {censored} model. This should be an unquoted column name
+#' although this argument is passed by expression and supports
 #' [quasiquotation][rlang::quasiquotation] (you can unquote column names). For
 #' `_vec()` functions, a numeric vector.
-#'
-#' @param eval_time The column identifier for the evaluation time points. This
-#' should be a numeric vector. This should be an unquoted column name although
-#' this argument is passed by expression and supports
-#' [quasiquotation][rlang::quasiquotation] (you can unquote column names). For
-#' `_vec()` functions, a numeric vector.
-#'
-#' @param ... Not currently used.
 #'
 #' @author Emil Hvitfeldt
 #'
@@ -56,9 +43,7 @@
 #' lung_surv %>%
 #'   brier_survival(
 #'     truth = surv_obj,
-#'     estimate = .pred_survival,
-#'     censoring_weights = ipcw,
-#'     eval_time = .time
+#'     .pred
 #'   )
 #' @export
 brier_survival <- function(data, ...) {
@@ -74,22 +59,15 @@ brier_survival <- new_dynamic_survival_metric(
 #' @export
 brier_survival.data.frame <- function(data,
                                       truth,
-                                      estimate,
-                                      censoring_weights,
-                                      eval_time,
+                                      ...,
                                       na_rm = TRUE,
-                                      case_weights = NULL,
-                                      ...) {
-  data <- dplyr::group_by(data, {{eval_time}})
-
+                                      case_weights = NULL) {
   dynamic_survival_metric_summarizer(
     name = "brier_survival",
     fn = brier_survival_vec,
     data = data,
     truth = !!enquo(truth),
-    estimate = !!enquo(estimate),
-    censoring_weights = !!enquo(censoring_weights),
-    eval_time = !!enquo(eval_time),
+    ...,
     na_rm = na_rm,
     case_weights = !!enquo(case_weights)
   )
@@ -99,43 +77,37 @@ brier_survival.data.frame <- function(data,
 #' @rdname brier_survival
 brier_survival_vec <- function(truth,
                                estimate,
-                               censoring_weights,
-                               eval_time,
                                na_rm = TRUE,
                                case_weights = NULL,
                                ...) {
   check_dynamic_survival_metric(
-    truth, estimate, censoring_weights, case_weights, eval_time
+    truth, estimate, case_weights
   )
-
-  n_distinct_time <- dplyr::n_distinct(eval_time)
-  if (n_distinct_time != 1) {
-    abort(paste0(
-      "`eval_time` should have at most 1 unique value. But ", n_distinct_time,
-      " were detected."
-    ))
-  }
-
   if (na_rm) {
     result <- yardstick_remove_missing(
-      truth, estimate, case_weights, censoring_weights, eval_time
+      truth, seq_along(estimate), case_weights
     )
 
     truth <- result$truth
-    estimate <- result$estimate
-    censoring_weights <- result$censoring_weights
-    eval_time <- result$eval_time
+    estimate <- estimate[result$estimate]
     case_weights <- result$case_weights
   } else {
     any_missing <- yardstick_any_missing(
-      truth, estimate, case_weights, censoring_weights, eval_time
+      truth, estimate, case_weights
     )
     if (any_missing) {
       return(NA_real_)
     }
   }
 
-  brier_survival_impl(truth, estimate, censoring_weights, case_weights, eval_time)
+  dplyr::tibble(estimate) %>%
+    tidyr::unnest(estimate) %>%
+    dplyr::group_by(.eval_time) %>%
+    dplyr::summarize(
+      .estimate = brier_survival_impl(
+        truth, .pred_survival, .weight_censored, case_weights, .eval_time
+      )
+    )
 }
 
 brier_survival_impl <- function(truth,
@@ -143,8 +115,8 @@ brier_survival_impl <- function(truth,
                                 censoring_weights,
                                 case_weights,
                                 eval_time) {
-  surv_time <- truth[, "time"]
-  surv_status <- truth[, "status"]
+  surv_time <- .extract_surv_time(truth)
+  surv_status <- .extract_surv_status(truth)
 
   if (!is.null(case_weights)) {
     norm_const <- sum(case_weights)