+++ R/surv-roc_survival_curve.R
@@ -10,17 +10,14 @@
 #' columns `.threshold`, `recall`, and `precision`.
 #'
 #' @seealso
-#' Compute the area under the ROC survival curve with (TODO link to
-#' roc_survival_auc()).
+#' Compute the area under the ROC survival curve with [roc_auc_survival()].
 #'
 #' @author Emil Hvitfeldt
 #' @examples
 #' result <- roc_curve_survival(
 #'   lung_surv,
 #'   truth = surv_obj,
-#'   estimate = .pred_survival,
-#'   censoring_weights = ipcw,
-#'   eval_time = .time
+#'   .pred
 #' )
 #' result
 #'
@@ -48,21 +45,16 @@ roc_curve_survival <- function(data, ...) {
 #' @rdname roc_curve_survival
 roc_curve_survival.data.frame <- function(data,
                                           truth,
-                                          estimate,
-                                          censoring_weights,
-                                          eval_time,
+                                          ...,
                                           na_rm = TRUE,
-                                          case_weights = NULL,
-                                          ...) {
+                                          case_weights = NULL) {
 
   result <- curve_survival_metric_summarizer(
     name = "roc_curve_survival",
     fn = roc_curve_survival_vec,
     data = data,
     truth = !!enquo(truth),
-    estimate = !!enquo(estimate),
-    censoring_weights = !!enquo(censoring_weights),
-    eval_time = !!enquo(eval_time),
+    ...,
     na_rm = na_rm,
     case_weights = !!enquo(case_weights)
   )
@@ -72,88 +64,79 @@ roc_curve_survival.data.frame <- function(data,
 
 roc_curve_survival_vec <- function(truth,
                                    estimate,
-                                   censoring_weights,
-                                   eval_time,
                                    na_rm = TRUE,
                                    case_weights = NULL,
                                    ...) {
   check_dynamic_survival_metric(
-    truth, estimate, censoring_weights, case_weights, eval_time
+    truth, estimate, case_weights
   )
 
   if (na_rm) {
     result <- yardstick_remove_missing(
-      truth, estimate, case_weights, censoring_weights, eval_time
+      truth, seq_along(estimate), case_weights
     )
 
     truth <- result$truth
-    estimate <- result$estimate
-    censoring_weights <- result$censoring_weights
-    eval_time <- result$eval_time
+    estimate <- estimate[result$estimate]
     case_weights <- result$case_weights
   } else {
     any_missing <- yardstick_any_missing(
-      truth, estimate, case_weights, censoring_weights, eval_time
+      truth, estimate, case_weights
     )
     if (any_missing) {
       return(NA_real_)
     }
   }
 
-  roc_curve_survival_impl(
-    truth = truth,
-    estimate = estimate,
-    censoring_weights = censoring_weights,
-    eval_time = eval_time
-  )
+  roc_curve_survival_impl(truth = truth, estimate = estimate)
 }
 
 roc_curve_survival_impl <- function(truth,
-                                    estimate,
-                                    censoring_weights,
-                                    eval_time) {
-  res <- dplyr::tibble(.threshold = sort(unique(c(0, 1, estimate))))
-  res$sensitivity <- vapply(
-    res$.threshold,
-    sensitivity_uno_2007,
-    FUN.VALUE = numeric(1),
-    eval_time, truth, estimate, censoring_weights
+                                    estimate) {
+  event_time <- .extract_surv_time(truth)
+  delta <- .extract_surv_status(truth)
+  data <- dplyr::tibble(event_time, delta, estimate)
+  data <- tidyr::unnest(data, cols = estimate)
+
+  res <- dplyr::tibble(.threshold = sort(unique(c(0, data$.pred_survival, 1))))
+
+  obs_time_le_time <- event_time <= data$.eval_time
+  obs_time_gt_time <- event_time > data$.eval_time
+  n <- nrow(data)
+  multiplier <- delta / (n * data$.weight_censored)
+
+  sensitivity_denom <- sum(obs_time_le_time * multiplier, na.rm = TRUE)
+  specificity_denom <- sum(obs_time_gt_time, na.rm = TRUE)
+
+  data_df <- data.frame(
+    le_time = obs_time_le_time,
+    ge_time = obs_time_gt_time,
+    multiplier = multiplier
   )
-  res$specificity <- vapply(
-    res$.threshold,
-    specificity_naive,
-    FUN.VALUE = numeric(1),
-    eval_time, truth, estimate
+  data_split <- split(data_df, data$.pred_survival)
+
+  sensitivity <- vapply(
+    data_split,
+    function(x) sum(x$le_time * x$multiplier, na.rm = TRUE),
+    FUN.VALUE = numeric(1)
   )
+  sensitivity <- cumsum(sensitivity)
+  sensitivity <- sensitivity / sensitivity_denom
+  sensitivity <- c(0, sensitivity, 1)
+  res$sensitivity <- sensitivity
+
+  specificity <- vapply(
+    data_split,
+    function(x) sum(x$ge_time, na.rm = TRUE),
+    FUN.VALUE = numeric(1)
+  )
+  specificity <- cumsum(specificity)
+  specificity <- specificity / specificity_denom
+  specificity <- c(0, specificity, 1)
+  specificity <- 1 - specificity
+  res$specificity <- specificity
+
   res
-}
-
-sensitivity_uno_2007 <- function(threshold,
-                                 eval_time,
-                                 surv_obj,
-                                 prob_surv,
-                                 prob_cens) {
-  n <- length(prob_surv)
-  event_time <- .extract_surv_time(surv_obj)
-  delta <- .extract_surv_status(surv_obj)
-  obs_time_le_time <- ifelse(event_time <= eval_time, 1, 0)
-  # Since the "marker" X is the survival prob, X <= C means an event
-  prob_le_thresh <- ifelse(prob_surv <= threshold, 1, 0)
-  multiplier <- delta / (n * prob_cens)
-  numer <- sum(obs_time_le_time * prob_le_thresh * multiplier, na.rm = TRUE)
-  denom <- sum(obs_time_le_time * multiplier, na.rm = TRUE)
-  numer / denom
-}
-
-specificity_naive <- function(threshold, eval_time, surv_obj, prob_surv) {
-  event_time <- .extract_surv_time(surv_obj)
-  delta <- .extract_surv_status(surv_obj)
-  obs_time_gt_time <- ifelse(event_time > eval_time, 1, 0)
-  # Since the "marker" X is the survival prob, X > C means no event
-  prob_gt_thresh <- ifelse(prob_surv > threshold, 1, 0)
-  numer <- sum(obs_time_gt_time * prob_gt_thresh, na.rm = TRUE)
-  denom <- sum(obs_time_gt_time, na.rm = TRUE)
-  numer / denom
 }
 
 # Dynamically exported