+++ tests/testthat/test-handle_missings.R
@@ -4,9 +4,7 @@ test_that("yardstick_remove_missing works", {
     list(
       truth = 1:10,
       estimate = 1:10,
-      case_weights = 1:10,
-      censoring_weights = NULL,
-      eval_time = NULL
+      case_weights = 1:10
     )
   )
 
@@ -15,9 +13,7 @@ test_that("yardstick_remove_missing works", {
     list(
       truth = c(1:4, 7:10),
       estimate = c(1:4, 7:10),
-      case_weights = c(1:4, 7:10),
-      censoring_weights = NULL,
-      eval_time = NULL
+      case_weights = c(1:4, 7:10)
     )
   )
 
@@ -26,9 +22,7 @@ test_that("yardstick_remove_missing works", {
     list(
       truth = c(1:4, 7:10),
       estimate = c(1:4, 7:10),
-      case_weights = c(1:4, 7:10),
-      censoring_weights = NULL,
-      eval_time = NULL
+      case_weights = c(1:4, 7:10)
     )
   )
 
@@ -37,9 +31,7 @@ test_that("yardstick_remove_missing works", {
     list(
       truth = c(1:4, 7:10),
       estimate = c(1:4, 7:10),
-      case_weights = c(1:4, 7:10),
-      censoring_weights = NULL,
-      eval_time = NULL
+      case_weights = c(1:4, 7:10)
     )
   )
 
@@ -48,9 +40,7 @@ test_that("yardstick_remove_missing works", {
     list(
       truth = c(1:4, 7:10),
       estimate = c(1:4, 7:10),
-      case_weights = c(1:4, 7:10),
-      censoring_weights = NULL,
-      eval_time = NULL
+      case_weights = c(1:4, 7:10)
     )
   )
 
@@ -59,39 +49,7 @@ test_that("yardstick_remove_missing works", {
     list(
       truth = c(2:4, 7:9),
       estimate = c(2:4, 7:9),
-      case_weights = c(2:4, 7:9),
-      censoring_weights = NULL,
-      eval_time = NULL
-    )
-  )
-
-  expect_identical(
-    yardstick_remove_missing(
-      c(NA, 2:10), c(1:9, NA), c(1:4, NA, NA, 7:10), c(NA, NA, 3:10)
-    ),
-    list(
-      truth = c(3:4, 7:9),
-      estimate = c(3:4, 7:9),
-      case_weights = c(3:4, 7:9),
-      censoring_weights = c(3:4, 7:9),
-      eval_time = NULL
-    )
-  )
-
-  expect_identical(
-    yardstick_remove_missing(
-      truth = c(NA, 2:10),
-      estimate = c(1:9, NA),
-      case_weights = c(1:4, NA, NA, 7:10),
-      censoring_weights = c(NA, NA, 3:10),
-      eval_time = c(1:2, NA, 4:10)
-    ),
-    list(
-      truth = c(4L, 7:9),
-      estimate = c(4L, 7:9),
-      case_weights = c(4L, 7:9),
-      censoring_weights = c(4L, 7:9),
-      eval_time = c(4L, 7:9)
+      case_weights = c(2:4, 7:9)
     )
   )
 })
@@ -99,14 +57,6 @@ test_that("yardstick_remove_missing works", {
 test_that("yardstick_any_missing works", {
   expect_false(
     yardstick_any_missing(1:10, 1:10, 1:10)
-  )
-
-  expect_false(
-    yardstick_any_missing(1:10, 1:10, 1:10, 1:10)
-  )
-
-  expect_false(
-    yardstick_any_missing(1:10, 1:10, 1:10, 1:10, 1:10)
   )
 
   expect_true(
@@ -127,13 +77,5 @@ test_that("yardstick_any_missing works", {
 
   expect_true(
     yardstick_any_missing(c(NA, 2:10), c(1:9, NA), c(1:4, NA, NA, 7:10))
-  )
-
-  expect_true(
-    yardstick_any_missing(1:10, 1:10, 1:10, c(NA, 2:10))
-  )
-
-  expect_true(
-    yardstick_any_missing(1:10, 1:10, 1:10, 1:10, c(NA, 2:10))
   )
 })